requirements.txt
# Deep Research Multi-Agent System Dependencies

# ========================================
# 🤖 AI AGENT FRAMEWORK
# ========================================
agno>=1.8.0
parallel-web>=0.2.0
openai>=1.0.0

# ========================================
# 🌐 WEB FRAMEWORK & API
# ========================================
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# ========================================
# 📊 DATA & VALIDATION
# ========================================
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ========================================
# 🌐 HTTP & NETWORKING
# ========================================
httpx>=0.25.0
aiohttp>=3.8.0
requests>=2.31.0

# ========================================
# ⚙️ CONFIGURATION & ENVIRONMENT
# ========================================
python-dotenv>=1.0.0

# ========================================
# 🗄️ DATABASE & STORAGE
# ========================================
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=4.5.0

# ========================================
# 🧪 TESTING & DEVELOPMENT
# ========================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# ========================================
# 📝 DOCUMENTATION & UTILITIES
# ========================================
rich>=13.0.0
typer>=0.9.0

# ========================================
# 🔧 DEVELOPMENT TOOLS (Optional)
# ========================================
# Uncomment for development
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
# pre-commit>=3.3.0
